backend/requirements.txt
appdirs==1.4.4
asgiref==3.4.1
autopep8==1.5.6
certifi==2020.12.5
click==7.1.2
coverage==5.5
distlib==0.3.1
Django==4.0.1
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.10
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysql-connector-python==8.0.26
mysqlclient==2.1.0
numpy==1.20.1
pandas==1.2.3
pep8==1.7.1
pipenv==2020.11.15
pycodestyle==2.7.0
pycparser==2.20
PyMySQL==1.0.2
pyserial==3.5
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
SQLAlchemy==1.4.37
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
virtualenv==20.4.3
virtualenv-clone==0.5.4
Werkzeug==1.0.1
